fixed error in data_faker.py
